Relocate `pregroup_tree` module
