+++ lambeq/text2diagram/pregroup_tree.py
