+++ tests/backend/test_grammar.py
@@ -2,7 +2,7 @@ from pytest import raises
 
 from lambeq import BobcatParser, SpacyTokeniser
 from lambeq.backend.grammar import *
-from lambeq.backend.pregroup_tree import PregroupTreeNode
+from lambeq.text2diagram.pregroup_tree import PregroupTreeNode
 
 
 def test_Ty():
